% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicMixWrapper.R
\name{cal_score_sparse}
\alias{cal_score_sparse}
\title{calculate the similarity score of two sparse matrices, explained in the SFAmix paper}
\usage{
cal_score_sparse(m1, m2, precis = FALSE)
}
\arguments{
\item{m1}{sparse matrix 1}

\item{m2}{sparse matrix 2}

\item{precis}{when TRUE, only use the most correlated components in the two
matrices to calculate the score, this is to account for the scenario where
the two matrices have different dimensions, but all components in the
smaller matrix have high correlation with a subset of big matrix}
}
\value{
a similarity score
}
\description{
calculate the similarity score of two sparse matrices, explained in the SFAmix paper
}
