% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicMixWrapper.R
\name{gen_SFA_data}
\alias{gen_SFA_data}
\title{Simulate matrix with dimension of 1000 x 200. Number of loadings and factors is set to 30, where 20 loadings and 20 factors are sparse. The sparse loadings and factors cotain mostly zeros, and random blocks of nonzero values generated from N(0,std). The dense loadings and factors are also generated from N(0,std). The error matrix is generated from N(0,1).}
\usage{
gen_SFA_data(
  std = 2,
  rsd = 123,
  std.err = 1,
  n.effects = 15,
  nfs = 10,
  ng = 1000,
  ns = 200,
  dense = T
)
}
\arguments{
\item{std}{standard deviation for the normal distribution of the non-zero entries of the sparse components}
}
\value{
a list containing the following

lams: the sparse loadings

lamd: the dense loadings

lam: the loading matrix combining both the sparse and dense loading

exs: the sparse factors matrix

exd: the dense factors matrix

ex: the factors matrix combining both the sparse and dense factors

y: the y matrix calculated as y = lam * ex + err
}
\description{
Simulate matrix with dimension of 1000 x 200. Number of loadings and factors is set to 30, where 20 loadings and 20 factors are sparse. The sparse loadings and factors cotain mostly zeros, and random blocks of nonzero values generated from N(0,std). The dense loadings and factors are also generated from N(0,std). The error matrix is generated from N(0,1).
}
