% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicMixWrapper.R
\name{gen_BicMix_data}
\alias{gen_BicMix_data}
\title{Simulate a matrix y = lam * ex + err}
\usage{
gen_BicMix_data(
  std = 2,
  rsd = NULL,
  std.err = 1,
  nf = 15,
  nfs = 10,
  ng = 500,
  ns = 200,
  type.loading = "mixture",
  type.factor = "dense"
)
}
\arguments{
\item{std}{standard deviation for the normal distribution of the non-zero entries of the sparse loading}

\item{rsd}{random seed}

\item{std.err}{standard deviation of the error term}

\item{nf}{total number of the loadings}

\item{nfs}{number of the sparse loadings}

\item{ng}{number of genes}

\item{ns}{number of samples}

\item{type.loading}{type of loading matrix wanted, "mixture" for a matrix mixed with sparse and dense components, "sparse" for a matrix with only sparse components}

\item{type.factor}{type of factor matrix wanted, "mixture" for a matrix mixed with sparse and dense components, "dense" for a matrix with only dense components}
}
\value{
a list containing the following

lams: the sparse loadings

lamd: the dense loadings

lam: the loading matrix combining both the sparse and dense loading

ex: the factors matrix combining both the sparse and dense factors

err: matrix of the error term

y: the y matrix calculated as y = lam * ex + err
}
\description{
Simulate a matrix y = lam * ex + err
}
