% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicMixWrapper.R
\name{gen_BicMix_data}
\alias{gen_BicMix_data}
\title{Simulate matrix with dimension of 500 x 20. Number of loadings and factors is set to 15, where 10 loadings and 10 factors are sparse. The sparse loadings and factors cotain mostly zeros, and random blocks of nonzero values generated from N(0,std). The dense loadings and factors are also generated from N(0,std). The error matrix is generated from N(0,1).}
\usage{
gen_BicMix_data(std = 2)
}
\arguments{
\item{std}{standard deviation for the normal distribution of the non-zero entries of the sparse components}
}
\value{
a list containing the following

lam: the loading matrix combining both the sparse and dense loading

ex: the factors matrix

y: the y matrix calculated as y = lam * ex + err
}
\description{
Simulate matrix with dimension of 500 x 20. Number of loadings and factors is set to 15, where 10 loadings and 10 factors are sparse. The sparse loadings and factors cotain mostly zeros, and random blocks of nonzero values generated from N(0,std). The dense loadings and factors are also generated from N(0,std). The error matrix is generated from N(0,1).
}
